Simple README description
Assuming this will change a bit, but I also included that we need to figure out a way to manage subject ID's across delayed condition (currently isn't a way to do that).  I think JATOS generates a new id each time, so that's still pending.  Everything else is just a basic overview of both tasks and clarifying any potentially confusing components, in case this goes up on OSF etc.